- CDMASystem 	- buildChannel: no channel is dismissed ("while(false)") - many more int -> uint - ChannelOrderEstimationSystem::ChannelOrderEstimationSystem 	- changed the implementation to look for the index of the maximum among the channel orders as well as the true one (old implementation, relying on int indexes, still there)

+++ algorithms/DSISoptAlgorithm.cpp
@@ -64,7 +64,7 @@ void DSISoptAlgorithm::process(const MatrixXd &observations, vector< double > no
 				_alphabet.int2symbolsArray(iTestedVector,testedVector);
 
 				// current tested vector is copied in the m-th position
-				for(k=0;k<static_cast<uint>(_nInputs);k++)
+				for(k=0;k<_nInputs;k++)
 					smoothingSymbolVectors(k,_channelOrder-1) = testedVector[k];
 
 				likelihoods(iTestedVector) = 0.0;
